optimized ChannelFilter and moved to own file
new file for ChannelFilter class implementation
use bitwise operations to reduce allocations
filter image already when updated from provider instead at paint time
